% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_session.R
\name{is_sessioninfo}
\alias{is_sessioninfo}
\title{Is identical session information object?}
\usage{
is_sessioninfo(x)
}
\arguments{
\item{x}{session information}
}
\description{
Is identical session information object?
}
